backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
pydantic==2.5.0
pydantic[email]
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
